add support for product and work same path for ftp add a test for subdirectories in work files fix a bug ignoring subdirectories in work files for ftp adapter

+++ tests/test_external_adapters.py
@@ -2,6 +2,7 @@ from pulse.api import *
 import unittest
 import os
 import mysql.connector as mariadb
+import ftplib
 
 test_dir = os.path.dirname(__file__)
 db = os.path.join(test_dir, "DB")
@@ -11,7 +12,7 @@ repos = os.path.join(test_dir, "repos")
 test_project_name = "testProj"
 
 
-# you have to set this a mysql database first, pulse user needs to create and drop database
+# you have to set this a mysql database first, pulse user needs rights to create and drop database
 db_host = "192.168.1.2"
 db_port = "3306"
 db_user = "pulse"
@@ -25,12 +26,36 @@ ftp_password = "okds-ki_se*84877sEE"
 ftp_root = "/pulseTest/"
 
 
+def ftp_rmtree(ftp, path):
+    """Recursively delete a directory tree on a remote server."""
+
+    wd = ftp.pwd()
+
+    try:
+        names = ftp.nlst(path)
+    except ftplib.all_errors as e:
+        return
+
+    for name in names:
+        if os.path.split(name)[1] in ('.', '..'):
+            continue
+        try:
+            ftp.cwd(name)  # if we can cwd to it, it's a folder
+            ftp.cwd(wd)  # don't try a nuke a folder we're in
+            ftp_rmtree(ftp, name)
+        except ftplib.all_errors:
+            ftp.delete(name)
+    try:
+        ftp.rmd(path)
+    except ftplib.all_errors as e:
+        print('FtpRmTree: Could not remove {0}: {1}'.format(path, e))
+
+
 def reset_files():
     cnx = mariadb.connect(host=db_host, port=db_port, user=db_user, password=db_password)
-    try:
+    if cnx.cursor().execute("SHOW DATABASES LIKE '" + test_project_name + "'"):
         cnx.cursor().execute("DROP DATABASE " + test_project_name)
-    except mariadb.DatabaseError:
-        pass
+        cnx.cmd_quit()
 
     for directory in [db, user_products, user_works, repos]:
         for path, subdirs, files in os.walk(directory):
@@ -48,16 +73,34 @@ def reset_files():
             except Exception as e:
                 print('Failed to delete %s. Reason: %s' % (file_path, e))
 
+    # clean ftp files
+    connection = ftplib.FTP()
+    connection.connect(ftp_host, ftp_port)
+    connection.login(ftp_login, ftp_password)
+    connection.cwd(ftp_root)
+    for project in connection.nlst():
+        if project.startswith("test"):
+            ftp_rmtree(connection, project)
+    # connection.close()
+    connection.quit()
+
     print "FILES RESET"
 
 
-def create_test_project(prj_name="test"):
+def create_test_project(prj_name=test_project_name):
     cnx = Connection({"DB_root": db})
     prj = cnx.create_project(
         prj_name,
         user_works,
         user_products,
-        default_repository_parameters={"root": os.path.join(repos, "default")}
+        default_repository_type="ftp_repo",
+        default_repository_parameters={
+            'host': ftp_host,
+            'port': ftp_port,
+            'login': ftp_login,
+            'password': ftp_password,
+            'root': ftp_root
+        }
     )
     create_template(prj, "mdl")
     create_template(prj, "surfacing")
@@ -82,12 +125,8 @@ class TestBasic(unittest.TestCase):
 
     def test_multiple_repository_types(self):
         cnx, prj = create_test_project()
-        prj.cfg.add_repository("serverB", "ftp_repo", {
-            'host': ftp_host,
-            'port': ftp_port,
-            'login': ftp_login,
-            'password': ftp_password,
-            'root': ftp_root
+        prj.cfg.add_repository("serverB", "shell_repo", {
+            'root': os.path.join(repos, "default")
         })
 
         template_resource = prj.create_template("rig", repository="serverB")
@@ -97,15 +136,15 @@ class TestBasic(unittest.TestCase):
         # self.assertTrue(os.path.exists(os.path.join(user_works, "test\\rig\\_template")))
         template_work.trash()
         prj.purge_unused_user_products()
-        self.assertFalse(os.path.exists(os.path.join(user_works, "test\\rig\\_template")))
-        # TODO : check out an uncomitted resource should not raise an error
+        self.assertFalse(os.path.exists(os.path.join(user_works, test_project_name, "rig\\_template")))
         template_resource.checkout()
-        self.assertTrue(os.path.exists(os.path.join(user_works, "test\\rig\\_template")))
+        self.assertTrue(os.path.exists(os.path.join(user_works, test_project_name, "rig\\_template")))
 
         # test moving resource between repo
+        self.assertTrue(os.path.exists(os.path.join(repos, "default", test_project_name, "work\\rig\\_template")))
         template_resource.set_repository("default")
 
-        self.assertTrue(os.path.exists(os.path.join(repos, "default\\test\\work\\rig\\_template")))
+        self.assertFalse(os.path.exists(os.path.join(repos, "default", test_project_name, "work\\rig\\_template")))
         # test moving resource between repo when the resource is locked
         template_resource.set_lock(True, "another_user")
         with self.assertRaises(PulseError):
@@ -163,6 +202,26 @@ class TestBasic(unittest.TestCase):
         rig2 = prj.get_resource("ch_anna", "rigging")
         self.assertTrue(rig2.get_commit("last").products[0] == 'actor_anim')
 
+    def test_work_subdirectories_are_commit(self):
+        subdirectory_name = "subdirtest"
+        # create a connection
+        cnx, prj = create_test_project()
+        # create a new template resource
+        create_template(prj, "modeling")
+        # create a resource based on this template
+        anna_mdl_resource = prj.create_resource("ch_anna", "modeling")
+        self.assertEqual(anna_mdl_resource.last_version, 0)
+
+        # checkout, and check directories are created
+        anna_mdl_work = anna_mdl_resource.checkout()
+        work_subdir_path = os.path.join(anna_mdl_work.directory, subdirectory_name)
+        os.makedirs(work_subdir_path)
+        open(work_subdir_path + "\\subdir_file.txt", 'a').close()
+        anna_mdl_work.commit()
+        anna_mdl_work.trash()
+        self.assertFalse(os.path.exists(anna_mdl_work.directory))
+        anna_mdl_resource.checkout()
+        self.assertTrue(os.path.exists(work_subdir_path + "\\subdir_file.txt"))
 
 if __name__ == '__main__':
     unittest.main()
